fix(padding): fix list mutation in loop
Previously, when the correct candidate came second in the list, a
single byte was left in the TCP buffer, which was then read in the
subsequent calls to validated_candidates as if it was a response to a
submitted IV. This caused a kind of off-by-one-error, where multiple
valid candidates were detected for other bytes. Thanks to an assertion
in the code, this was detected, leading to an exception instead of
propagating the error and leading to incorrect results.

test(padding): Add three failure cases from previous test runs.